requirements.txt
groq
langgraph
langchain-groq
langchain-core
python-dotenv
SpeechRecognition
pyaudio
pyttsx3
pinecone
sentence-transformers
numpy